requirements.txt
streamlit>=1.29.0
groq>=0.4.0
pandas>=2.0.0
python-dotenv>=1.0.0
